requirements.txt
# Core dependencies
graphviz>=0.20.1
pillow>=10.0.0
customtkinter>=5.2.0
cairosvg>=2.7.1

# XML processing
lxml>=4.9.0

# Excel processing
openpyxl>=3.1.2

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0